feat(handler): makes all types of readers have the same way processing cancel-related events
